6 Charts working
