+++ SocialDashboard/app/assets/javascripts/charts/bubbleChart/bubbleChart.js
@@ -1,0 +1,94 @@
+
+var displayBubbleChart;
+var ready = function() {
+    displayBubbleChart = function(list, div) {
+        var bubbleChart = new d3.svg.BubbleChart({
+            supportResponsive: true,
+            //container: => use @default
+            size: 600,
+            //viewBoxSize: => use @default
+            innerRadius: 600 / 3.5,
+            //outerRadius: => use @default
+            radiusMin: 50,
+            //radiusMax: use @default
+            //intersectDelta: use @default
+            //intersectInc: use @default
+            //circleColor: use @default
+            data: {
+                items: list,
+                eval: function (item) {return item.count;},
+                classed: function (item) {return item.text.split(" ").join("");}
+            },
+            plugins: [
+                {
+                    name: "central-click",
+                    options: {
+                        text: "",
+                        style: {
+                            "font-size": "12px",
+                            "font-style": "italic",
+                            "font-family": "Source Sans Pro, sans-serif",
+                            //"font-weight": "700",
+                            "text-anchor": "middle",
+                            "fill": "white"
+                        },
+                        attr: {dy: "65px"},
+                        centralClick: function(a) {
+                            console.log(a);
+                        }
+                    }
+                },
+                {
+                    name: "lines",
+                    options: {
+                        format: [
+                            {// Line #0
+                                textField: "count",
+                                classed: {count: true},
+                                style: {
+                                    "font-size": "28px",
+                                    "font-family": "Source Sans Pro, sans-serif",
+                                    "text-anchor": "middle",
+                                    fill: "white"
+                                },
+                                attr: {
+                                    dy: "0px",
+                                    x: function (d) {return d.cx;},
+                                    y: function (d) {return d.cy;}
+                                }
+                            },
+                            {// Line #1
+                                textField: "text",
+                                classed: {text: true},
+                                style: {
+                                    "font-size": "10px",
+                                    "font-family": "Source Sans Pro, sans-serif",
+                                    "text-anchor": "middle",
+                                    fill: "white"
+                                },
+                                attr: {
+                                    dy: "20px",
+                                    x: function (d) {return d.cx;},
+                                    y: function (d) {return d.cy;}
+                                }
+                            }
+                        ],
+                        centralFormat: [
+                            {// Line #0
+                                style: {"font-size": "50px"},
+                                attr: {}
+                            },
+                            {// Line #1
+                                style: {"font-size": "20px"},
+                                attr: {dy: "40px"}
+                            }
+                        ]
+                    }
+                }]
+        }, div);
+    }
+
+}
+
+$(document).on('page:load', ready);
+$(document).ready(ready);
