+++ SocialDashboard/app/assets/stylesheets/mainStyles.css
@@ -213,3 +213,11 @@ html, body {
     color: #d1d1d1 !important;
 }
 
+/*Styles for the loading spinner*/
+#loadingDiv {
+    width: 100%;
+    height: 100%;
+    background: url('icons/ajax-loader.gif') no-repeat center center;
+    z-index: 100;
+}
+
