+++ SocialDashboard/app/controllers/charts_controller.rb
@@ -33,8 +33,6 @@ class ChartsController < ApplicationController
 
     def get_cookie_filter(json_filter, type)
       data = JSON.parse(json_filter)
-      p json_filter
-      p type
       if type.eql?("trends") || type.eql?("popular_terms")
         filter = PhrasesFilter.new(data['language'], generate_country_list(data['country_list']), data['start_date'], data['end_date'],data['type'])
       else
